requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and tokenizers
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0
accelerate>=0.24.0

# Distributed training
deepspeed>=0.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Monitoring
wandb>=0.15.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
psutil>=5.9.0
gputil>=1.4.0
py-cpuinfo>=9.0.0

# Code quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
